refactor: Split monolithic tester into modular structure
- Add tester/ module with separate concerns:
  - ssh_client.py: SSH connection management
  - claim_extractor.py: Parse docs, extract claims
  - executor.py: Execute commands/actions
  - validator.py: Validate results
  - reporter.py: Generate reports
- Add config.py: VM connection details
- Add main.py: CLI entry point
- Remove old monolithic tester.py
