+++ docs_tester/config.py
@@ -1,0 +1,7 @@
+"""VM Connection Configuration"""
+
+VM_HOST = "10.10.30.140"
+VM_USER = "cm"
+VM_PASSWORD = "123"
+NETDATA_URL = "http://10.10.30.140:19999"
+OUTPUT_DIR = "test_results"
