+++ docs_tester/tester/__init__.py
@@ -1,0 +1,18 @@
+"""Netdata Documentation Tester"""
+
+from .ssh_client import SSHClient
+from .claim_extractor import ClaimExtractor, StepType, Step, Workflow
+from .executor import Executor
+from .validator import Validator
+from .reporter import Reporter
+
+__all__ = [
+    'SSHClient',
+    'ClaimExtractor',
+    'Executor',
+    'Validator',
+    'Reporter',
+    'StepType',
+    'Step',
+    'Workflow'
+]
